requirements.txt
# It is important to install torch, torchvision, and torchaudio together with compatible versions.
# For torch==2.1.0, the compatible versions are torchvision==0.16.0 and torchaudio==2.1.0.
# Downgrading transformers to avoid the torch.load security check for now.
# Pinning versions of diffusers, huggingface-hub and tokenizers to resolve dependency conflicts.
# Pinning pyarrow to a version compatible with datasets<=2.13.0
# Adding bitsandbytes for 8-bit optimizer to save memory.

torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
diffusers==0.21.4
transformers==4.34.1
huggingface-hub==0.17.3
tokenizers==0.14.1
Pillow==10.2.0
modelscope==1.9.5
pyarrow==12.0.1
bitsandbytes==0.41.1
tqdm==4.66.2
scipy==1.12.0
numpy==1.26.4
